change to float

+++ src/diffusers/pipelines/controlnet/pipeline_controlnet_sd_xl_img2img.py
@@ -272,7 +272,7 @@ class StableDiffusionXLControlNetImg2ImgPipeline(
         prompt: str,
         prompt_2: Optional[str] = None,
         device: Optional[torch.device] = None,
-        cfg_end: Optional[int] = None,
+        cfg_end: Optional[float] = None,
         num_images_per_prompt: int = 1,
         do_classifier_free_guidance: bool = True,
         negative_prompt: Optional[str] = None,
